test(hir-ty): cover public type interface
