+++ crates/hir-ty/src/type_infer.rs
@@ -857,323 +857,3 @@ fn stmt_of(db: &dyn TyDb, stmt: InContainer<hir_def::stmt::StmtId>) -> Option<hi
         }
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use std::fmt;
-
-    use base_db::{
-        diagnostics_config::DiagnosticsConfig,
-        project::{CompilationProfile, CompilationProfileId, PreprocessConfig, ProjectConfig},
-        salsa::{self, Durability},
-        source_db::{
-            FileLoader, SourceDb, SourceDbStorage, SourceFileKind, SourceRootDb,
-            SourceRootDbStorage,
-        },
-        source_root::{SourceRoot, SourceRootId},
-    };
-    use hir_def::{
-        db::{HirDefDb, HirDefDbStorage, InternDbStorage},
-        module::ModuleId,
-        symbol::{DefOriginLoc, NameContext},
-    };
-    use preproc_expand::db::PreprocDbStorage;
-    use rustc_hash::FxHashSet;
-    use smol_str::SmolStr;
-    use triomphe::Arc;
-    use utils::paths::{AbsPathBuf, Utf8PathBuf};
-    use vfs::{AnchoredPath, FileId, FileSet, VfsPath};
-
-    use super::*;
-    use crate::{TypeSystem, db::TyDbStorage, display::HirDisplay};
-
-    const TOP: FileId = FileId::from_raw(0);
-    const ROOT: SourceRootId = SourceRootId(0);
-    const PROFILE: CompilationProfileId = CompilationProfileId(0);
-
-    #[salsa::database(
-        SourceDbStorage,
-        SourceRootDbStorage,
-        PreprocDbStorage,
-        InternDbStorage,
-        HirDefDbStorage,
-        TyDbStorage
-    )]
-    #[derive(Default)]
-    struct TestDb {
-        storage: salsa::Storage<Self>,
-    }
-
-    impl salsa::Database for TestDb {}
-
-    impl fmt::Debug for TestDb {
-        fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-            f.debug_struct("TestDb").finish()
-        }
-    }
-
-    impl FileLoader for TestDb {
-        fn resolve_path(&self, path: AnchoredPath<'_>) -> Option<FileId> {
-            let source_root_id = SourceRootDb::source_root_id(self, path.anchor);
-            SourceRootDb::source_root(self, source_root_id).resolve_path(path)
-        }
-    }
-
-    fn db_with_root_text(root_text: &str) -> TestDb {
-        let top_path = abs_path("rtl/top.sv");
-        let mut file_set = FileSet::default();
-        file_set.insert(TOP, VfsPath::from(top_path.clone()));
-        let root = SourceRoot::new_local_with_source_files(file_set, vec![TOP]);
-        let mut files = FxHashSet::default();
-        files.insert(TOP);
-
-        let preprocess = PreprocessConfig::default();
-        let project_config = ProjectConfig::new(
-            vec![Some(PROFILE)],
-            vec![CompilationProfile {
-                source_roots: vec![ROOT],
-                top_modules: Vec::new(),
-                preprocess: preprocess.clone(),
-            }],
-        );
-
-        let mut db = TestDb::default();
-        db.set_files_with_durability(Box::new(files), Durability::HIGH);
-        db.set_project_config_with_durability(Arc::new(project_config), Durability::HIGH);
-        db.set_diagnostics_config_with_durability(
-            Arc::new(DiagnosticsConfig::default()),
-            Durability::HIGH,
-        );
-        db.set_source_root_with_durability(ROOT, Arc::new(root), Durability::LOW);
-        db.set_source_root_id_with_durability(TOP, ROOT, Durability::LOW);
-        db.set_file_path_with_durability(TOP, Some(top_path), Durability::LOW);
-        db.set_file_kind_with_durability(TOP, SourceFileKind::SystemVerilog, Durability::LOW);
-        db.set_file_text_with_durability(TOP, Arc::from(root_text), Durability::LOW);
-        db
-    }
-
-    fn abs_path(path: &str) -> AbsPathBuf {
-        let prefix = if cfg!(windows) { "C:/repo" } else { "/repo" };
-        AbsPathBuf::assert(Utf8PathBuf::from(format!("{prefix}/{path}")))
-    }
-
-    fn ident(name: &str) -> Ident {
-        SmolStr::new(name)
-    }
-
-    fn module_id(db: &TestDb, name: &str) -> ModuleId {
-        db.unit_scope()
-            .module_ids(db, &ident(name))
-            .unique()
-            .expect("module should resolve uniquely")
-    }
-
-    fn decl_ty(db: &TestDb, module_id: ModuleId, name: &str) -> Ty {
-        let res = resolve_name(db, module_id.into(), &ident(name), NameContext::Value);
-        let decl = res
-            .iter()
-            .find_map(|def_id| def_id.primary_origin(db).as_decl(db))
-            .expect("resolved value should include a declaration");
-        TypeSystem::new(db).type_of_def(DefId::new(db, decl)).ty().clone()
-    }
-
-    fn path_ty(db: &TestDb, module_id: ModuleId, segments: &[&str]) -> Ty {
-        let path = segments.iter().map(|segment| ident(segment)).collect::<Vec<_>>();
-        let res = hir_def::pathres::resolve_path(db, module_id.into(), &path, NameContext::Value);
-        assert!(!res.is_unresolved(), "path {segments:?} should resolve");
-        TypeSystem::new(db).type_of_resolution(res).ty().clone()
-    }
-
-    fn typedef_ty(db: &TestDb, module_id: ModuleId, name: &str) -> Ty {
-        let res = resolve_name(db, module_id.into(), &ident(name), NameContext::Type);
-        let typedef = res
-            .iter()
-            .find_map(|def_id| def_id.primary_origin(db).as_typedef(db))
-            .expect("resolved type should include a typedef");
-        TypeSystem::new(db).type_of_def(DefId::new(db, typedef)).ty().clone()
-    }
-
-    fn assert_owner_is_decl(db: &TestDb, def: DefId, name: &str) {
-        let DefOriginLoc::Decl(decl) = def.primary_origin(db).loc(db) else {
-            panic!("expected {name} owner to be a declaration");
-        };
-        assert_eq!(decl.cont_id.data(db).declarator(decl.value).name.as_deref(), Some(name));
-    }
-
-    fn assert_owner_is_typedef(db: &TestDb, def: DefId, name: &str) {
-        let DefOriginLoc::Typedef(typedef) = def.primary_origin(db).loc(db) else {
-            panic!("expected {name} owner to be a typedef");
-        };
-        assert_eq!(typedef.cont_id.data(db).typedef(typedef.value).name.as_deref(), Some(name));
-    }
-
-    #[test]
-    fn sv_type_adt_covers_enum_union_and_unpacked_array_kinds() {
-        let db = db_with_root_text(
-            r#"
-module m;
-  typedef enum logic [1:0] { A, B } state_t;
-  enum { C, D } anon_enum;
-  typedef union packed { logic [7:0] byte_v; int int_v; } payload_u;
-  logic queue_var[$];
-  logic bounded_queue[$:4];
-  logic assoc_var[string];
-  logic dyn_var[];
-  event ev;
-  chandle handle;
-endmodule
-"#,
-        );
-        let module_id = module_id(&db, "m");
-
-        let Ty::Alias { target: state_target, .. } = typedef_ty(&db, module_id, "state_t") else {
-            panic!("state_t should infer as an alias");
-        };
-        let Ty::Enum(state_def) = *state_target else {
-            panic!("state_t target should be an enum");
-        };
-        assert_owner_is_typedef(&db, state_def, "state_t");
-
-        let Ty::Enum(anon_enum_def) = decl_ty(&db, module_id, "anon_enum") else {
-            panic!("anonymous enum declaration should infer as Ty::Enum");
-        };
-        assert_owner_is_decl(&db, anon_enum_def, "anon_enum");
-
-        let Ty::Alias { target: payload_target, .. } = typedef_ty(&db, module_id, "payload_u")
-        else {
-            panic!("payload_u should infer as an alias");
-        };
-        let Ty::Union(payload_def) = *payload_target else {
-            panic!("payload_u target should be a union");
-        };
-        assert_owner_is_typedef(&db, payload_def, "payload_u");
-
-        assert!(matches!(decl_ty(&db, module_id, "queue_var"), Ty::Queue { size: None, .. }));
-        assert!(matches!(
-            decl_ty(&db, module_id, "bounded_queue"),
-            Ty::Queue { size: Some(_), .. }
-        ));
-        assert!(matches!(decl_ty(&db, module_id, "assoc_var"), Ty::Assoc { .. }));
-        assert!(matches!(decl_ty(&db, module_id, "dyn_var"), Ty::Dynamic(_)));
-        assert!(matches!(decl_ty(&db, module_id, "ev"), Ty::Event));
-        assert!(matches!(decl_ty(&db, module_id, "handle"), Ty::Chandle));
-    }
-
-    #[test]
-    fn sv_type_adt_display_covers_new_variants() {
-        let db = db_with_root_text(
-            r#"
-module m;
-  typedef enum { A, B } state_t;
-  typedef union packed { logic [7:0] byte_v; int int_v; } payload_u;
-  logic queue_var[$];
-  logic bounded_queue[$:4];
-  logic assoc_var[string];
-  logic dyn_var[];
-  event ev;
-  chandle handle;
-endmodule
-"#,
-        );
-        let module_id = module_id(&db, "m");
-
-        let rendered = [
-            typedef_ty(&db, module_id, "state_t").display_source(&db).unwrap(),
-            typedef_ty(&db, module_id, "payload_u").display_source(&db).unwrap(),
-            decl_ty(&db, module_id, "queue_var").display_source(&db).unwrap(),
-            decl_ty(&db, module_id, "bounded_queue").display_source(&db).unwrap(),
-            decl_ty(&db, module_id, "assoc_var").display_source(&db).unwrap(),
-            decl_ty(&db, module_id, "dyn_var").display_source(&db).unwrap(),
-            decl_ty(&db, module_id, "ev").display_source(&db).unwrap(),
-            decl_ty(&db, module_id, "handle").display_source(&db).unwrap(),
-        ];
-
-        assert_eq!(
-            rendered,
-            [
-                "state_t",
-                "payload_u",
-                "logic [$]",
-                "logic [$:4]",
-                "logic [string]",
-                "logic []",
-                "event",
-                "chandle",
-            ]
-        );
-    }
-
-    #[test]
-    fn virtual_interface_type_display_covers_instance_and_modport() {
-        let db = db_with_root_text(
-            r#"
-interface bus_if;
-  wire clk;
-  modport host(input clk);
-endinterface
-
-module top;
-  bus_if u_if();
-endmodule
-"#,
-        );
-        let top = module_id(&db, "top");
-
-        assert_eq!(
-            path_ty(&db, top, &["u_if"]).display_source(&db).unwrap(),
-            "virtual interface bus_if"
-        );
-        assert_eq!(
-            path_ty(&db, top, &["u_if", "host"]).display_source(&db).unwrap(),
-            "virtual interface bus_if.host"
-        );
-    }
-
-    #[test]
-    fn program_instance_type_displays_as_module_shaped_definition() {
-        let db = db_with_root_text(
-            r#"
-program p;
-endprogram
-
-module top;
-  p u_p();
-endmodule
-"#,
-        );
-        let top = module_id(&db, "top");
-        let program = module_id(&db, "p");
-
-        let program_res = Resolution::Unique(DefId::new(&db, program));
-        assert_eq!(
-            TypeSystem::new(&db)
-                .display_source(&TypeSystem::new(&db).type_of_resolution(program_res))
-                .unwrap(),
-            "p"
-        );
-        assert_eq!(path_ty(&db, top, &["u_p"]).display_source(&db).unwrap(), "p");
-    }
-
-    #[test]
-    fn streaming_with_range_display_preserves_with_keyword() {
-        let db = db_with_root_text(
-            r#"
-module m(input logic [3:0] a);
-  logic [3:0] x = {<<{a with [3:0]}};
-endmodule
-"#,
-        );
-        let module_id = module_id(&db, "m");
-        let module = db.module(module_id);
-        let (stream_id, _) = module
-            .exprs
-            .iter()
-            .find(|(_, expr)| matches!(expr, hir_def::expr::Expr::Stream { .. }))
-            .expect("streaming concatenation should lower");
-
-        assert_eq!(
-            InContainer::new(module_id.into(), stream_id).display_source(&db).unwrap(),
-            "{<<{a with [3:0]}}"
-        );
-    }
-}
